+++ R/boot_ci.R
@@ -97,7 +97,7 @@
 #' boot_p(bs$gender)
 #' summary(fit)$coefficients[3, ]
 #'
-#'
+#' \dontrun{
 #' # 'spread_coef()' from the 'sjmisc'-package makes it easy to generate
 #' # bootstrapped statistics like confidence intervals or p-values
 #' library(dplyr)
@@ -125,7 +125,7 @@
 #'   # spread model coefficient for all 100 models
 #'   spread_coef(models, append = FALSE) %>%
 #'   # compute the CI for all bootstrapped model coefficients
-#'   boot_ci()
+#'   boot_ci()}
 #'
 #' @importFrom stats qt quantile
 #' @importFrom dplyr quos